Revert "[tkt-13] Reorganise project structure"
